% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_cold_wave.R
\name{coldwave_draw_anorm}
\alias{coldwave_draw_anorm}
\title{draw_anorm}
\usage{
coldwave_draw_anorm(
  data_anorm,
  time_beg = make_datetime(2023, 8, 1),
  time_end = make_datetime(2023, 11, 15)
)
}
\arguments{
\item{data_anorm}{data.table with columns of \code{date}, \code{mean}, \code{TEM}}
}
\description{
draw_anorm
}
