% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_cold_wave.R
\name{coldwave_draw_DRT}
\alias{coldwave_draw_DRT}
\title{draw_DRT}
\usage{
coldwave_draw_DRT(
  dat_hour,
  time_beg = make_datetime(2023, 8, 1),
  time_end = make_datetime(2023, 11, 15),
  ...,
  lims_Tair = c(5, 40),
  lims_DRT = c(0, 16)
)
}
\arguments{
\item{dat_hour}{data.table with columns of \code{date}, \code{TEM}}
}
\description{
draw_DRT
}
