% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_cold_wave.R
\name{coldwave_draw_main}
\alias{coldwave_draw_main}
\title{coldwave_draw_main}
\usage{
coldwave_draw_main(
  dat_day,
  clim,
  time_beg = make_datetime(2023, 8, 1),
  time_end = make_datetime(2023, 11, 15)
)
}
\arguments{
\item{dat_day}{data.table with columns: date, TEM}

\item{clim}{data table, with columns: date, variable, mean, sd}
}
\description{
coldwave_draw_main
}
