% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVT.R
\name{IVT}
\alias{IVT}
\title{Integrated Vapor Transport}
\usage{
IVT(uwind, vwind, shum, levs)
}
\arguments{
\item{uwind, vwind}{horizontal and vertical wind (m/s), in the dimension of \verb{[nlat, nlon, nlev, ntime]}}

\item{shum}{specific humidity (kg/kg), in the same dimension of \verb{[nlat, nlon, nlev, ntime]}}

\item{levs}{air pressure (hPa), a numberic vector with the length of \code{nvel}}
}
\value{
IVT (kg m-1 s-1)
\itemize{
\item IVT_u: horizontal
\item IVT_t: vertical
}
}
\description{
Integrated Vapor Transport
}
