% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ivt.R
\name{plot_ivt}
\alias{plot_ivt}
\title{ggplot_ivt}
\usage{
plot_ivt(
  df,
  uv_scalar = 0.04,
  title = NULL,
  arrow.len = 0.3,
  arrow.size = 0.2,
  outfile = NULL
)
}
\arguments{
\item{df}{with the column of \code{lon}, \code{lat}, \code{u}, \code{v} and \code{wind}}

\item{uv_scalar}{global Environment: worldmap}
}
\description{
ggplot_ivt
}
