% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CUG.met-package.R
\docType{package}
\name{CUG.met-package}
\alias{CUG.met}
\alias{CUG.met-package}
\title{CUG.met: MeteoroHydrology in the CUG}
\description{
MeteoroHydrology in the CUG.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
